requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.1
Flask-APScheduler==1.11.0
gmpy2==2.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
krakenex==2.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.25.3
patsy==0.5.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
regex==2019.12.20
requests==2.22.0
scikit-learn==0.22
scipy==1.3.2
seaborn==0.9.0
selenium==3.141.0
simplejson==3.17.0
six==1.12.0
soupsieve==1.9.5
statsmodels==0.10.1
tornado==6.0.3
tqdm==4.32.1
tzlocal==2.0.0
urllib3==1.25.7
virtualenv==16.7.4
Werkzeug==1.0.0
wincertstore==0.2
yahoo-finance==1.4.0
